Add routing
* Router module with matcher macros
* Commander module with [Nadia] sender macros
* You may define command handlers in other modules
