+++ lib/app/commander.ex
@@ -9,157 +9,7 @@ defmodule App.Commander do
       import App.Commander
       alias Nadia.Model
       alias Nadia.Model.InlineQueryResult
-
-      def match_message(message) do
-        try do
-          apply __MODULE__, :do_match_message, [message]
-        rescue
-          err in FunctionClauseError ->
-            Logger.log :warn, """
-              Errored when matching command. #{Poison.encode! err}
-              Message was: #{Poison.encode! message}
-              """
-        end
-      end
     end
-  end
-
-  def generate_message_matcher(function) do
-    quote do
-      def do_match_message(var!(update)) do
-        Task.async fn -> unquote(function) end
-      end
-    end
-  end
-
-  defp generate_command(command, function) do
-    quote do
-      def do_match_message(%{
-        message: %{
-          text: "/" <> unquote(command)
-        }
-      } = var!(update)) do
-        Task.async fn -> unquote(function) end
-      end
-
-      def do_match_message(%{
-        message: %{
-          text: "/" <> unquote(command) <> " " <> _
-        }
-      } = var!(update)) do
-        Task.async fn -> unquote(function) end
-      end
-
-      def do_match_message(%{
-        message: %{
-          text: "/" <> unquote(command) <> "@" <> unquote(@bot_name)
-        }
-      } = var!(update)) do
-        Task.async fn -> unquote(function) end
-      end
-
-      def do_match_message(%{
-        message: %{
-          text: "/" <> unquote(command) <> "@" <> unquote(@bot_name) <> " " <> _
-        }
-      } = var!(update)) do
-        Task.async fn -> unquote(function) end
-      end
-    end
-  end
-
-  def generate_inline_query_matcher(function) do
-    quote do
-      def do_match_message(%{inline_query: inline_query} = var!(update))
-      when not is_nil(inline_query) do
-        Task.async fn -> unquote(function) end
-      end
-    end
-  end
-
-  def generate_inline_query_command(command, function) do
-    quote do
-      def do_match_message(%{
-        inline_query: %{query: "/" <> unquote(command)}
-      } = var!(update)) do
-        Task.async fn -> unquote(function) end
-      end
-
-      def do_match_message(%{
-        inline_query: %{query: "/" <> unquote(command) <> " " <> _}
-      } = var!(update)) do
-        Task.async fn -> unquote(function) end
-      end
-    end
-  end
-
-  def generate_callback_query_matcher(function) do
-    quote do
-      def do_match_message(%{callback_query: callback_query} = var!(update))
-      when not is_nil(callback_query) do
-        Task.async fn -> unquote(function) end
-      end
-    end
-  end
-
-  def generate_callback_query_command(command, function) do
-    quote do
-      def do_match_message(%{
-        callback_query: %{data: "/" <> unquote(command)}
-      } = var!(update)) do
-        Task.async fn -> unquote(function) end
-      end
-
-      def do_match_message(%{
-        callback_query: %{data: "/" <> unquote(command) <> " " <> _}
-      } = var!(update)) do
-        Task.async fn -> unquote(function) end
-      end
-    end
-  end
-
-  # Receiver Macros
-
-  defmacro message(do: function) do
-    generate_message_matcher(function)
-  end
-
-  defmacro command(commands, do: function)
-  when is_list(commands) do
-    Enum.map commands, fn command ->
-      generate_command(command, function)
-    end
-  end
-  defmacro command(command, do: function) do
-    generate_command(command, function)
-  end
-
-  defmacro inline_query(do: function) do
-    generate_inline_query_matcher(function)
-  end
-
-  defmacro inline_query_command(commands, do: function)
-  when is_list(commands) do
-    Enum.map commands, fn item ->
-      generate_inline_query_command(item, function)
-    end
-  end
-  defmacro inline_query_command(command, do: function) do
-    generate_inline_query_command(command, function)
-  end
-
-  defmacro callback_query(do: function) do
-    generate_callback_query_matcher(function)
-  end
-
-  defmacro callback_query_command(commands, do: function)
-  when is_list(commands) do
-    Enum.map commands, fn item ->
-      generate_callback_query_command(item, function)
-    end
-  end
-  defmacro callback_query_command(command, do: function) do
-    generate_callback_query_command(command, function)
   end
 
   # Sender Macros
